# Optimization: Make hashing over transactions snarks partial (ledger hashes)