--- a/src/lib/mina_state/snarked_ledger_state.ml
+++ b/src/lib/mina_state/snarked_ledger_state.ml
@@ -199,6 +199,48 @@ module Make_str (A : Wire_types.Concrete) = struct
       [@@deriving compare, equal, hash, sexp, yojson]
 
       let to_latest = Fn.id
+
+      (* `work_id` is computed by hashing the full statement *)
+
+      let work_id = hash
+
+      (* Override hash functions with faster partial versions.
+       * The assumption here is that the combination of source and target ledger
+       * hashes is unique enough to differentiate between statements, so
+       * there will be no clashes.
+       *)
+
+      let hash_fold_t state (t : t) =
+        let { source =
+                { first_pass_ledger = source_first_pass_ledger
+                ; second_pass_ledger = source_second_pass_ledger
+                ; pending_coinbase_stack = _
+                ; local_state = _
+                }
+            ; target =
+                { first_pass_ledger = target_first_pass_ledger
+                ; second_pass_ledger = target_second_pass_ledger
+                ; pending_coinbase_stack = _
+                ; local_state = _
+                }
+            ; connecting_ledger_left = _
+            ; connecting_ledger_right = _
+            ; supply_increase = _
+            ; fee_excess = _
+            ; sok_digest = ()
+            } =
+          t
+        in
+        let hash_fold_t hash state =
+          Frozen_ledger_hash.Stable.V1.hash_fold_t state hash
+        in
+        state
+        |> hash_fold_t source_first_pass_ledger
+        |> hash_fold_t source_second_pass_ledger
+        |> hash_fold_t target_first_pass_ledger
+        |> hash_fold_t target_second_pass_ledger
+
+      let hash = Ppx_hash_lib.Std.Hash.of_fold hash_fold_t
     end
   end]
 
@@ -288,6 +330,8 @@ module Make_str (A : Wire_types.Concrete) = struct
     input
 
   let to_field_elements t = Random_oracle.pack_input (to_input t)
+
+  let work_id = Stable.V2.work_id
 
   module Checked = struct
     type t = var