--- a/src/lib/mina_state/snarked_ledger_state_intf.ml
+++ b/src/lib/mina_state/snarked_ledger_state_intf.ml
@@ -182,6 +182,8 @@ module type Full = sig
 
   val to_field_elements : t -> Tick.Field.t array
 
+  val work_id : t -> int
+
   module Checked : sig
     type t = var