@tizoc 2023-02-16
@mrmr1993 thanks! I will adjust accordingly (I also need to investigate what changed, I see that now the ledger is changed in multiple passes).

> For example, I think that it's possible to have a coinbase 'internal command' with a zero balance, which will result in no change to the ledgers (@deepthiskumar can you confirm?).

I see, I assumed there was always going to be a coinbase transaction affecting the ledger (and thus changing the hash).