@deepthiskumar 2023-02-21
> For example, I think that it's possible to have a coinbase 'internal command' with a zero balance, which will result in no change to the ledgers (@deepthiskumar can you confirm?). 

Coinbase amount is set by the network config and added during block validation and so if a block producer [says there's a coinbase transaction in this block](https://github.com/MinaProtocol/mina/blob/04ce62015ec23dcff42b25f0c4ec7ec44b0b0116/src/lib/staged_ledger_diff/diff.ml#L139) then the amount will be whatever is set in the network config. A block could have no coinbase transaction though but that doesn't matter in this context.
Fee transfers (prover fee, txns fee) could be zero but in the current out of snark implementation, only non-zero amount fee transfers are created. Again, 
Re, failed transactions with zero transaction fee, there'd still be a ledger change due to nonce increment, no?