@tizoc 2023-02-21
> Re, failed transactions with zero transaction fee, there'd still be a ledger change due to nonce increment, no?

This is one of the main assumptions made in this change, even failed transactions will result in a ledger update (either because of a fee or the nonce). But when it comes to coinbase transactions and zkapp transactions there may be corner cases that I didn't consider.

But stepping back a bit. There can be `hash` clashes without issues (since these are OCaml hashes used by data structures and this is expected to happen anyway). Hash sets and hash tables will still work fine (the key value will be compared to disambiguate).

The worry is more about avoiding work-id clashes, because `work_id = hash`. These can still happen with full hashing, but partial hashing could increase the chances.

One option here (if we are not 100% sure that the fields used to compute the hash are enough) is to use partial hashing (using the source and target ledger hashes for both phases) for the OCaml hash, but to not use the partial hash as work-id. Instead keep hashing the full statement to compute the work-id by not using `hash` here:

```ocaml
  let work_ids t : int One_or_two.t =
    One_or_two.map t ~f:Transaction_snark.Statement.hash
```

and add a new `Transaction_snark.Statement.work_id` function.

What do you think? are there any downsides to that approach?