@deepthiskumar 2023-02-21
Coinbase or fee transfers don't do nonce increments so if we allowed zero amount coinbases or fee transfers then this would be a problem but currently the impl doesn't allow it. Applying a zkapp transaction will result in nonce increment for the fee payer part at the very least.

>But stepping back a bit. There can be hash clashes without issues (since these are OCaml hashes used by data structures and this is expected to happen anyway). Hash sets and hash tables will still work fine (the key value will be compared to disambiguate).

There is still a problem when retrieving values since you only get the most recently added one? and so having clashes could result in invalid blocks whenever this occurs. Proofs retrieved could be for different statements and they are not verified again during block production.

>to not use the partial hash as work-id.

Not changing the work id computation makes sense. Work ids are mainly used for debugging, to track a piece of snark work through daemon and snark workers. It doesn't impact block validation at all, I think.