@tizoc 2023-02-21
> Coinbase or fee transfers don't do nonce increments so if we allowed zero amount coinbases or fee transfers then this would be a problem but currently the impl doesn't allow it. Applying a zkapp transaction will result in nonce increment for the fee payer part at the very least.

Great, thanks for clarifying.

>There is still a problem when retrieving values since you only get the most recently added one? and so having clashes could result in invalid blocks whenever this occurs. Proofs retrieved could be for different statements and they are not verified again during block production.

It will still work fine because here only the `hash` function is being modified, the `compare` function remains the same, and the values used as keys also remain the same.

Because the key value still needs to be compared to ensure that the key is really the same (and not just that the integer hash matches), hash clashes will not be an issue:

```ocaml
utop # #require "core";;
utop # open! Core;;
utop # 
  module StringSameHash = struct
      type t = string
      let compare = String.compare
      let hash_fold_t state _  = String.hash_fold_t state ""
      let hash _ = String.hash ""
      let sexp_of_t _ = failwith "unimplemented"
  end
;;
(* output ->
  module StringSameHash :
  sig
    type t = string
    val compare : t -> t -> int
    val hash_fold_t :
      Base_internalhash_types.state -> 'a -> Base_internalhash_types.state
    val hash : 'a -> int
    val sexp_of_t : 'a -> 'b
  end *)
utop # StringSameHash.hash "a", StringSameHash.hash "b";;
- : int * int = (0, 0)
utop #
let samehash_table = Hashtbl.create (module StringSameHash) in
ignore @@ Hashtbl.add samehash_table ~key:"a" ~data:1;
ignore @@ Hashtbl.add samehash_table ~key:"b" ~data:2;
(Hashtbl.find samehash_table "a", Hashtbl.find samehash_table "b");;
- : int option * int option = (Some 1, Some 2)
```