@tizoc 2023-02-21
Just pushed a new version rebased (and adjusted) over current develop. It also uses the full hashing functions to compute `work_id`.