@tizoc 2023-02-21
Note that I ignored the connecting ledger hashes and only used the source + target ledger hashes of both passes, but if needed, I can add those too.