@tizoc 2023-07-19
Not necessary anymore because of https://github.com/MinaProtocol/mina/pull/13409